// lib/types.ts
export interface Customer {
  _id: string;
  name: string;
  email: string;
  phone?: string;
  totalSpending: number;
  lastVisit?: string;
  createdAt: string;
  updatedAt: string;
}